Use float instead of int
min_time is 0.5-1 second, so we can't round it

+++ flask_core.py
@@ -241,11 +241,11 @@ def solvechallenge():
     if challenge_details is None:
         return jsonify({'success': False, 'error': 'Invalid challenge_id'})
 
-    if int(challenge_details['expires']) < time.time():
+    if float(challenge_details['expires']) < time.time():
         return jsonify({'success': False, 'error': 'Challenge expired'})
 
     try:
-        if int(challenge_details['min_time']) > time.time():
+        if float(challenge_details['min_time']) > time.time():
             db_connection.delete(challenge_id)
             return jsonify({'success': False, 'error': 'Wrong answer'})
 
@@ -294,7 +294,7 @@ def verify():
     if token_details is None:
         return jsonify({'success': False, 'error': 'Response token is invalid'})
 
-    if int(token_details['expires']) < time.time():
+    if float(token_details['expires']) < time.time():
         return jsonify({'success': False, 'error': 'Response token expired'})
 
 # we delete tokens immediately once they are used
